+++ src/components/add-collection-project-pop/styles.styl
@@ -1,0 +1,9 @@
+@import '../constants.styl'
+
+.projectResult
+  background-color: primary-background
+  &.private
+    background-color: private
+  
+.resultWrap
+  display: flex
